Set ALLOWED_HOSTS to ["*"] in settings.py
This change updates the ALLOWED_HOSTS setting in settings.py to ["*"],
allowing all hosts to access the application. Previously, no hosts were
allowed. This change is necessary to enable access to the application
from any domain or IP address. It is particularly useful during
development and testing phases. No issues are associated with this
commit.